feat: display LoginModal on click add account
