+++ web-marketplace/src/components/organisms/LoginButton/LoginButton.tsx
@@ -1,23 +1,19 @@
 import { useQuery } from '@tanstack/react-query';
 import { REGEN_DENOM } from 'config/allowedBaseDenoms';
-import { useSetAtom } from 'jotai';
-import { postData } from 'utils/fetch/postData';
 
 import OutlinedButton from 'web-components/lib/components/buttons/OutlinedButton';
 import { EmailConfirmationModal } from 'web-components/lib/components/modal/EmailConfirmationModal/EmailConfirmationModal';
 
 import { useLedger } from 'ledger';
-import { apiUri } from 'lib/apiUri';
-import { errorBannerTextAtom } from 'lib/atoms/error.atoms';
 import { getBalanceQuery } from 'lib/queries/react-query/cosmos/bank/getBalanceQuery/getBalanceQuery';
-import { getCsrfTokenQuery } from 'lib/queries/react-query/registry-server/getCsrfTokenQuery/getCsrfTokenQuery';
 
 import { useAuthData } from 'hooks/useAuthData';
 
 import { chainId } from '../../../lib/ledger';
 import { useWallet } from '../../../lib/wallet/wallet';
+import { useEmailConfirmationData } from '../LoginFlow/hooks/useEmailConfirmationData';
+import { LoginFlow } from '../LoginFlow/LoginFlow';
 import { LoginModal } from '../LoginModal/LoginModal';
-import { useEmailConfirmationData } from './hooks/useEmailConfirmationData';
 import { useLoginData } from './hooks/useLoginData';
 import {
   EMAIL_CONFIRMATION_CANCEL,
@@ -36,30 +32,18 @@ type Props = {
 
 const LoginButton = ({ size = 'small' }: Props) => {
   const styles = useLoginButtonStyles();
-  const { wallet, walletConnectUri } = useWallet();
-  const {
-    isConfirmationModalOpen,
-    email,
-    emailModalError,
-    onConfirmationModalClose,
-    onMailCodeChange,
-    onResendPasscode,
-    setIsConfirmationModalOpen,
-    setEmail,
-  } = useEmailConfirmationData();
-
+  const { wallet } = useWallet();
   const {
     connecting,
     isModalOpen,
     modalState,
-    onButtonClick,
     onModalClose,
     qrCodeUri,
     isWaitingForSigning,
     setIsWaitingForSigningAtom,
     walletsUiConfig,
+    onButtonClick,
   } = useLoginData();
-
   const { noAccountAndNoWallet } = useAuthData();
 
   const { bankClient } = useLedger();
@@ -73,9 +57,6 @@ const LoginButton = ({ size = 'small' }: Props) => {
     }),
   );
 
-  const { data: token } = useQuery(getCsrfTokenQuery({}));
-  const setErrorBannerTextAtom = useSetAtom(errorBannerTextAtom);
-
   return chainId ? (
     <>
       <div className={styles.root}>
@@ -86,59 +67,8 @@ const LoginButton = ({ size = 'small' }: Props) => {
             </OutlinedButton>
           )}
         </>
+        <LoginFlow />
       </div>
-      <LoginModal
-        open={isModalOpen}
-        onClose={onModalClose}
-        wallets={walletsUiConfig}
-        socialProviders={socialProviders}
-        onEmailSubmit={async ({ email }) => {
-          if (token) {
-            try {
-              setEmail(email);
-              await postData({
-                url: `${apiUri}/marketplace/v1/auth/passcode`,
-                data: {
-                  email,
-                },
-                token,
-              });
-              onModalClose();
-              setIsConfirmationModalOpen(true);
-            } catch (e) {
-              setErrorBannerTextAtom(String(e));
-            }
-          }
-        }}
-        state={modalState}
-        qrCodeUri={qrCodeUri}
-        connecting={connecting}
-      />
-      <EmailConfirmationModal
-        resendText={RESEND_TEXT}
-        resendButtonLink={{
-          text: RESEND_BUTTON_TEXT,
-          onClick: onResendPasscode,
-        }}
-        cancelButton={{
-          text: EMAIL_CONFIRMATION_CANCEL,
-          onClick: onConfirmationModalClose,
-        }}
-        signInButton={{
-          text: EMAIL_CONFIRMATION_SUBMIT,
-          disabled: true,
-          onClick: () => void 0,
-        }}
-        mailLink={{ text: email, href: '#' }}
-        onClose={onConfirmationModalClose}
-        open={isConfirmationModalOpen}
-        error={emailModalError}
-        onCodeChange={onMailCodeChange}
-      />
-      <MobileSigningModal
-        isOpen={isWaitingForSigning && !!walletConnectUri}
-        onClose={() => setIsWaitingForSigningAtom(false)}
-      />
     </>
   ) : (
     <></>
